requirements.txt
langchain-ollama
requests
beautifulsoup4
jinja2
weasyprint
pydantic
python-dotenv
pytest
